Moved all "ReadyTask" styles into a single file, but broke out the logic into ReadyTaskBallot, ReadyTaskPlan, and ReadyTaskRegister instead of trying to do the differences with if/else statements. Fixed some ESlint warnings.
