+++ src/js/components/Ballot/MeasureItem.jsx
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { withStyles, withTheme } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
 import { historyPush } from '../../utils/cordovaUtils';
 import { renderLog } from '../../utils/logging';
 import MeasureStore from '../../stores/MeasureStore';
@@ -14,7 +13,6 @@ import BallotItemSupportOpposeComment from '../Widgets/BallotItemSupportOpposeCo
 
 class MeasureItem extends Component {
   static propTypes = {
-    classes: PropTypes.object,
     measureWeVoteId: PropTypes.string.isRequired,
     // theme: PropTypes.object,
   };
@@ -86,7 +84,6 @@ class MeasureItem extends Component {
   render () {
     renderLog('MeasureItem');  // Set LOG_RENDER_EVENTS to log all renders
     // const { supportProps, transitioning } = this.state;
-    const { classes } = this.props;
     let {
       ballotItemDisplayName, stateDisplayName,
     } = this.state;
